requirements.txt
beautifulsoup4
langchain
langchain-core
langchain-google-genai
lxml
nltk
playwright
python-dotenv
requests
xhtml2pdf
